Small fixes to u8vectors and symbols after a bunch of debugging and learning how the GC works

+++ vm/include/heap.h
@@ -86,8 +86,8 @@ uint16 _SYS_RAM_TO_VEC_OBJ(uint16 o)
 #else
 #define OBJ_TO_RAM_ADDR(o,f) (((uint16)((o) - MIN_RAM_ENCODING) << 2) + (f))
 #define OBJ_TO_ROM_ADDR(o,f) (((uint16)((o) - MIN_ROM_ENCODING) << 2) + (CODE_START + 4 + (f)))
-#define VEC_TO_RAM_BASE_ADDR(o) (((uint16)(o) + MAX_RAM_ENCODING + 1 - MIN_RAM_ENCODING) << 2)
-#define VEC_TO_ROM_BASE_ADDR(o) (((uint16)(o) << 2) + (CODE_START + 4))
+#define VEC_TO_RAM_BASE_ADDR(o) ((uint16)(((uint16)(o) + MAX_RAM_ENCODING + 1 - MIN_RAM_ENCODING) << 2))
+#define VEC_TO_ROM_BASE_ADDR(o) ((uint16)(((uint16)(o) << 2) + (CODE_START + 4)))
 // don't use these with OBJ_TO_RAM_ADDR, danger!
 #define _SYS_VEC_TO_RAM_OBJ(o) ((o) + MAX_RAM_ENCODING + 1)
 #define _SYS_RAM_TO_VEC_OBJ(o) ((o) - MAX_RAM_ENCODING - 1)
